improved pythonization tests

+++ test/pythonizables.cxx
@@ -43,10 +43,10 @@ pyzables::SharedCountable_t pyzables::gime_mine() { return mine; }
 pyzables::SharedCountable_t* pyzables::gime_mine_ptr() { return &mine; }
 pyzables::SharedCountable_t& pyzables::gime_mine_ref() { return mine; }
 
-unsigned long pyzables::pass_mine_sp(std::shared_ptr<Countable> ptr) { return ptr->m_check; }
-unsigned long pyzables::pass_mine_sp_ref(std::shared_ptr<Countable>& ptr) { return ptr->m_check; }
-unsigned long pyzables::pass_mine_sp_ptr(std::shared_ptr<Countable>* ptr) { return (*ptr)->m_check; }
+unsigned int pyzables::pass_mine_sp(std::shared_ptr<Countable> ptr) { return ptr->m_check; }
+unsigned int pyzables::pass_mine_sp_ref(std::shared_ptr<Countable>& ptr) { return ptr->m_check; }
+unsigned int pyzables::pass_mine_sp_ptr(std::shared_ptr<Countable>* ptr) { return (*ptr)->m_check; }
 
-unsigned long pyzables::pass_mine_rp(Countable c) { return c.m_check; }
-unsigned long pyzables::pass_mine_rp_ref(const Countable& c) { return c.m_check; }
-unsigned long pyzables::pass_mine_rp_ptr(const Countable* c) { return c->m_check; }
+unsigned int pyzables::pass_mine_rp(Countable c) { return c.m_check; }
+unsigned int pyzables::pass_mine_rp_ref(const Countable& c) { return c.m_check; }
+unsigned int pyzables::pass_mine_rp_ptr(const Countable* c) { return c->m_check; }
